Move `model/plot` to `model/result/plot` as it's a part of Result
