+++ src/components/plot-list/index.tsx
@@ -4,7 +4,7 @@ import {ActionHandler} from '../../actions/redux-action';
 import {RESULT_LIMIT_INCREASE, ResultAction} from '../../actions/result';
 import {ShelfAction} from '../../actions/shelf';
 import {Bookmark} from '../../models/bookmark';
-import {PlotObject} from '../../models/plot';
+import {PlotObject} from '../../models/result';
 import {ResultType} from '../../models/result';
 import {Plot} from '../plot';
 import * as styles from './plot-list.scss';
