+++ src/models/result/plot.ts
@@ -9,8 +9,8 @@ import {ExtendedGroupBy} from 'compassql/build/src/query/groupby';
 import {toMap} from 'compassql/build/src/util';
 import {Data} from 'vega-lite/build/src/data';
 import {FacetedCompositeUnitSpec} from 'vega-lite/build/src/spec';
-import {fromFieldQuery} from './shelf';
-import {ShelfFieldDef} from './shelf/encoding';
+import {fromFieldQuery} from '../shelf';
+import {ShelfFieldDef} from '../shelf/encoding';
 
 export interface PlotFieldInfo {
   fieldDef: ShelfFieldDef;
